// workspaces/liferay-one-workspace/client-extensions/liferay-one-custom-element/src/types/salesforceContract.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export type SalesforceContract = {
	endDate?: string;
	externalReferenceCode: string;
	id: number;
	name?: string;
	startDate?: string;
};